Support control of an external amplifier.
Add a -A <script> option to integrate with an external amplifier. The
given script is called to save the amplifier state on playback start,
restore it on end and to handle all volume changes (the host's mixer
volume is set to max).

An example script is provided for a Samsung soundbar with Multiroom/WiFi
capability (specifically an HW-K850 which this feature was developed and
tested against).

An example systemd service file is also provided to show how to run
squeezelite on Linux as a daemon belonging to the display manager user.

Signed-off-by: Mike Jagdis <mjagdis@eris-associates.co.uk>