Merge pull request #117: Fix memory leak and double free with ffmpeg.
If avcodec_decode_audio4() fails to decode a frame, ff_decode() returns
without freeing the AVPacket. This can add up to a significant amount of
leaked memory when many decoding errors are encountered.

ff->readbuf, as passed to avio_alloc_context(), has the potential to be
freed and reallocated by libavio behind the scenes. This can and does
lead to a double free because the original reference, which is
no longer valid, is freed instead of the current allocation.

Thanks tofurky.

Increase squeezelite revision to 1294.