Reduce CPU usage during FLAC playback
The FLAC decoder will call read_cb() as quickly as possible until it
gets the data it needs to continue decoding, even if the callback is
returning 0 bytes.

Add a 1000us delay in read_cb if it's entered when no data is in the
stream buffer.

With this change, CPU usage drops from 100% to about 15% during playback
of an ogg/flac stream.

Note that the issue seems to be most prominent when using a larger than
default output buffer size - -b :16384 was enough to trigger it in
testing.