Make it possible to disable mp3 decoder(s) at compile time
OpenWrt has a patch[1] for squeezelite to disable the build of the mad
mp3 decoder. The reason for removing the mad decoder is probably to
avoid having to install two mp3 decoders (mad and mpg123) on devices
running OpenWrt (typically routers and other small devices with
limited storage).

I rewrote the patch to make it configurable and more generic (mad and
mpg123 decoders can both be disabled individually).

Do you think this is useful to squeezelite in general and worth
merging?

[1] https://github.com/openwrt/packages/blob/master/sound/squeezelite/patches/020-no_libmad.patch