Fix manual loading of OpenSSL 1.1
sslsym.c doesn't compile with OpenSSL 1.1.x because functions
SSLv23_client_method and SSL_library_init are deprecated and replaced
with macros in openssl/ssl.h.

And SSL_CTX_set_options was a macro in OpenSSL 1.0.x but is a function
in OpenSSL 1.1.x which needs to be dynamically loaded.

Compile and runtime tested with OpenSSL versions 1.0.1, 1.1.1n and 1.1.1q.