Fix premature end of tracks on Windows.
I think we have a final view and fix of the problem which is a general problem, not specific really to us.

I'll use an example to describe the problem, although it is not exactly the flow of events for squeezelite

- an mp3 file is 3.5 MB at 160 kbits/s = 20 kB/s
- a client reads it at the exact required speed, 20kB/sec, let's say one recv() of 20kB per second, no pre-buffering for simplicity
- some OS, like Windows, can have very large TCP kernel buffer (about 3MB or more) and with a fast connection, the TCP windows size is widely open
- in a matter of seconds, the whole file is sent to the client, let's say that about 3MB are in the kernel buffers
- as far as the server is concerned, all has been sent and acknowledge, so it does a close()
- the close() sends a FIN to the client which responds by an ACK and the server enters FIN_WAIT_2 state
- BUT, at that point from a client point of view, all recv() will have plenty of read for the next 150 s before it sees the eof!
- so client, will not do a close() and thus will not send a FIN
- the server is in FIN_WAIT_2 state and according to the TCP specs, it should stay like that *forever*
- now, various OS (Windows at least) start a timer similar to TIME_WAIT (120s) when starting a close(), or when receiving the ACK of their FIN, that I don't know (in fact Windows has a specific registry entry for that, see below). This is to more aggressively deal with half-closed sockets.
- of course, after 120s, the server has not seen a client's FIN and sends a RST
- that RST is received by the client and causes an error there and all the remaining data in the TCP buffers to be discarded and lost
- of course, none of that happens with high bitrate formats as the client consumes data fast enough so that the kernel TCP buffers are never idle for 120s and it might not happen for low bit rate when the application buffering system reads it all. It has to be the bad combination of bitrate, file size and kernel's buffers... hence it does not happen all the time.

The registry key to set FIN_WAIT_2is HKEY_LOCAL_MACHINE\SYSTEM\CurrentControlSet\Servic es\Tcpip\Parameters\TcpFinWait2Delay, a DWORD between 30 to 300s.

In Windows, the getsockopt of so_rcvbuffer returns 64kB and setting it with that returned value solves the problem. Why? It's because Windows uses a feature named TCP auto tuning where the buffers are adjusted automatically depending on estimation of BW/latency. Setting so_rcvbuf disables that feature on a per-socket level (It can be done at the system level as well).

-philippe44.

Thanks philippe44 for tracking down the cause of this issue.
Increase squeezelite revision to 1419.